requirements.txt
asgiref==3.5.0
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.4.4
Babel==2.10.3
backports.entry-points-selectable==1.1.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
dataclasses==0.6
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
dj-email-url==1.0.5
Django==2.2.13
django-allauth==0.51.0
django-babel==0.6.2
django-cache-url==3.4.2
django-countries==7.3.2
django-crispy-forms==1.14.0
django-htmx==1.8.0
django-phonenumber-field==6.3.0
django-rename-app==0.1.3
environs==9.5.0
execnet==1.9.0
filelock==3.3.1
idna==3.3
iniconfig==1.1.1
marshmallow==3.17.0
numpy==1.21.3
oauthlib==3.2.0
packaging==21.2
pandas==1.3.4
pep8==1.7.1
phonenumbers==8.12.51
phonenumberslite==8.12.51
Pillow==9.1.1
platformdirs==2.4.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.5.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cache==1.0
pytest-pep8==1.0.6
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2018.5
requests==2.28.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.2.4
toml==0.10.2
typing_extensions==4.3.0
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.9.0
xlrd==2.0.1
